feat: move data fetching to App.tsx
