+++ src/App.tsx
@@ -1,4 +1,8 @@
+import { useEffect, useState } from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+
+import { Product } from 'dummyjson-api/models'
+import { getProducts } from 'dummyjson-api'
 
 import PageLayout from 'components/PageLayout'
 import ProductsList from 'pages/ProductsList'
@@ -7,11 +11,23 @@ import Cart from 'pages/Cart'
 import NoMatch from 'pages/NoMatch'
 
 function App() {
+  const [products, setProducts] = useState<Product[]>([])
+
+  // TODO: move to the application root
+  useEffect(() => {
+    const fetchProducts = async () => {
+      const data = await getProducts()
+      setProducts(data.products)
+    }
+
+    fetchProducts()
+  }, [])
+
   return (
     <Router>
       <Routes>
         <Route path="/" element={<PageLayout />}>
-          <Route index element={<ProductsList />} />
+          <Route index element={<ProductsList products={products} />} />
           <Route path="product-details/:id" element={<ProductDetails />} />
           <Route path="cart" element={<Cart />} />
           <Route path="*" element={<NoMatch />} />
